src/python/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1 # Use [standard] for common extras like websockets, http-tools

# Pydantic (Core and Settings)
pydantic==2.6.1
pydantic-settings==2.1.0

# Supabase Client
supabase==2.4.1 # Or a more recent compatible version (check supabase-py releases)

# Langchain and OpenAI
langchain==0.1.5
langchain-openai==0.0.5
openai==1.12.0

# Utilities
python-dotenv==1.0.1
email-validator==2.1.0.post1 # Required by FastAPI for EmailStr
python-multipart==0.0.6 # For form data parsing if needed

# --- Testing Libraries ---
pytest>=8.0.0,<9.0.0 # Current major stable version of pytest
pytest-asyncio>=0.23.0,<0.24.0 # Handles async tests, compatible with pytest 8
httpx>=0.27.0,<0.28.0 # Async HTTP client used by FastAPI's TestClient and our tests
pytest-mock>=3.12.0,<4.0.0 # Provides the 'mocker' fixture